Reorganize input
