+++ crates/gpui/src/elements/mod.rs
@@ -5,7 +5,6 @@ mod deferred;
 mod div;
 mod image_cache;
 mod img;
-mod input;
 mod list;
 mod surface;
 mod svg;
@@ -20,7 +19,6 @@ pub use deferred::*;
 pub use div::*;
 pub use image_cache::*;
 pub use img::*;
-pub use input::*;
 pub use list::*;
 pub use surface::*;
 pub use svg::*;
