+++ crates/gpui/src/input.rs
@@ -1,190 +1,11 @@
-use crate::{App, Bounds, Context, Entity, InputHandler, Pixels, UTF16Selection, Window};
-use std::ops::Range;
+mod input;
+mod input_handler;
 
-/// Implement this trait to allow views to handle textual input when implementing an editor, field, etc.
-///
-/// Once your view implements this trait, you can use it to construct an [`ElementInputHandler<V>`].
-/// This input handler can then be assigned during paint by calling [`Window::handle_input`].
-///
-/// See [`InputHandler`] for details on how to implement each method.
-pub trait EntityInputHandler: 'static + Sized {
-    /// See [`InputHandler::text_for_range`] for details
-    fn text_for_range(
-        &mut self,
-        range: Range<usize>,
-        adjusted_range: &mut Option<Range<usize>>,
-        window: &mut Window,
-        cx: &mut Context<Self>,
-    ) -> Option<String>;
-
-    /// See [`InputHandler::selected_text_range`] for details
-    fn selected_text_range(
-        &mut self,
-        ignore_disabled_input: bool,
-        window: &mut Window,
-        cx: &mut Context<Self>,
-    ) -> Option<UTF16Selection>;
-
-    /// See [`InputHandler::marked_text_range`] for details
-    fn marked_text_range(
-        &self,
-        window: &mut Window,
-        cx: &mut Context<Self>,
-    ) -> Option<Range<usize>>;
-
-    /// See [`InputHandler::unmark_text`] for details
-    fn unmark_text(&mut self, window: &mut Window, cx: &mut Context<Self>);
-
-    /// See [`InputHandler::replace_text_in_range`] for details
-    fn replace_text_in_range(
-        &mut self,
-        range: Option<Range<usize>>,
-        text: &str,
-        window: &mut Window,
-        cx: &mut Context<Self>,
-    );
-
-    /// See [`InputHandler::replace_and_mark_text_in_range`] for details
-    fn replace_and_mark_text_in_range(
-        &mut self,
-        range: Option<Range<usize>>,
-        new_text: &str,
-        new_selected_range: Option<Range<usize>>,
-        window: &mut Window,
-        cx: &mut Context<Self>,
-    );
-
-    /// See [`InputHandler::bounds_for_range`] for details
-    fn bounds_for_range(
-        &mut self,
-        range_utf16: Range<usize>,
-        element_bounds: Bounds<Pixels>,
-        window: &mut Window,
-        cx: &mut Context<Self>,
-    ) -> Option<Bounds<Pixels>>;
-
-    /// See [`InputHandler::character_index_for_point`] for details
-    fn character_index_for_point(
-        &mut self,
-        point: crate::Point<Pixels>,
-        window: &mut Window,
-        cx: &mut Context<Self>,
-    ) -> Option<usize>;
-
-    /// See [`InputHandler::accepts_text_input`] for details
-    fn accepts_text_input(&self, _window: &mut Window, _cx: &mut Context<Self>) -> bool {
-        true
-    }
-}
-
-/// The canonical implementation of [`crate::PlatformInputHandler`]. Call [`Window::handle_input`]
-/// with an instance during your element's paint.
-pub struct ElementInputHandler<V> {
-    view: Entity<V>,
-    element_bounds: Bounds<Pixels>,
-}
-
-impl<V: 'static> ElementInputHandler<V> {
-    /// Used in [`Element::paint`][element_paint] with the element's bounds, a `Window`, and a `App` context.
-    ///
-    /// [element_paint]: crate::Element::paint
-    pub fn new(element_bounds: Bounds<Pixels>, view: Entity<V>) -> Self {
-        ElementInputHandler {
-            view,
-            element_bounds,
-        }
-    }
-}
-
-impl<V: EntityInputHandler> InputHandler for ElementInputHandler<V> {
-    fn selected_text_range(
-        &mut self,
-        ignore_disabled_input: bool,
-        window: &mut Window,
-        cx: &mut App,
-    ) -> Option<UTF16Selection> {
-        self.view.update(cx, |view, cx| {
-            view.selected_text_range(ignore_disabled_input, window, cx)
-        })
-    }
-
-    fn marked_text_range(&mut self, window: &mut Window, cx: &mut App) -> Option<Range<usize>> {
-        self.view
-            .update(cx, |view, cx| view.marked_text_range(window, cx))
-    }
-
-    fn text_for_range(
-        &mut self,
-        range_utf16: Range<usize>,
-        adjusted_range: &mut Option<Range<usize>>,
-        window: &mut Window,
-        cx: &mut App,
-    ) -> Option<String> {
-        self.view.update(cx, |view, cx| {
-            view.text_for_range(range_utf16, adjusted_range, window, cx)
-        })
-    }
-
-    fn replace_text_in_range(
-        &mut self,
-        replacement_range: Option<Range<usize>>,
-        text: &str,
-        window: &mut Window,
-        cx: &mut App,
-    ) {
-        self.view.update(cx, |view, cx| {
-            view.replace_text_in_range(replacement_range, text, window, cx)
-        });
-    }
-
-    fn replace_and_mark_text_in_range(
-        &mut self,
-        range_utf16: Option<Range<usize>>,
-        new_text: &str,
-        new_selected_range: Option<Range<usize>>,
-        window: &mut Window,
-        cx: &mut App,
-    ) {
-        self.view.update(cx, |view, cx| {
-            view.replace_and_mark_text_in_range(
-                range_utf16,
-                new_text,
-                new_selected_range,
-                window,
-                cx,
-            )
-        });
-    }
-
-    fn unmark_text(&mut self, window: &mut Window, cx: &mut App) {
-        self.view
-            .update(cx, |view, cx| view.unmark_text(window, cx));
-    }
-
-    fn bounds_for_range(
-        &mut self,
-        range_utf16: Range<usize>,
-        window: &mut Window,
-        cx: &mut App,
-    ) -> Option<Bounds<Pixels>> {
-        self.view.update(cx, |view, cx| {
-            view.bounds_for_range(range_utf16, self.element_bounds, window, cx)
-        })
-    }
-
-    fn character_index_for_point(
-        &mut self,
-        point: crate::Point<Pixels>,
-        window: &mut Window,
-        cx: &mut App,
-    ) -> Option<usize> {
-        self.view.update(cx, |view, cx| {
-            view.character_index_for_point(point, window, cx)
-        })
-    }
-
-    fn accepts_text_input(&mut self, window: &mut Window, cx: &mut App) -> bool {
-        self.view
-            .update(cx, |view, cx| view.accepts_text_input(window, cx))
-    }
-}
+pub use input::{Input, InputLineLayout};
+// todo: move to keymap
+pub use input::{
+    Backspace, Copy, Cut, Delete, Down, End, Enter, Home, Left, MoveToBeginning, MoveToEnd, Paste,
+    Right, SelectAll, SelectDown, SelectLeft, SelectRight, SelectToBeginning, SelectToEnd,
+    SelectUp, SelectWordLeft, SelectWordRight, Tab, Up, WordLeft, WordRight,
+};
+pub use input_handler::*;
